Premier refacto des controllers

+++ src/Controller/Admin/MediaController.php
@@ -4,6 +4,7 @@ namespace App\Controller\Admin;
 
 use App\Entity\Media;
 use App\Form\MediaType;
+use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -13,9 +14,17 @@ use Symfony\Component\Routing\Attribute\Route;
 
 class MediaController extends AbstractController
 {
+
+	public function __construct(
+		private readonly EntityManagerInterface $entityManager,
+	)
+	{
+	}
+
 	#[Route(path: '/admin/media', name: 'admin_media_index')]
 	public function index(Request $request): Response
 	{
+		$mediasRepo = $this->entityManager->getRepository(Media::class);
         $page = $request->query->getInt('page', 1);
 
         $criteria = [];
@@ -24,13 +33,13 @@ class MediaController extends AbstractController
             $criteria['user'] = $this->getUser();
         }
 
-        $medias = $this->getDoctrine()->getRepository(Media::class)->findBy(
+        $medias = $mediasRepo->findBy(
             $criteria,
             ['id' => 'ASC'],
             25,
             25 * ($page - 1)
         );
-        $total = $this->getDoctrine()->getRepository(Media::class)->count([]);
+        $total = $mediasRepo->count();
 
         return $this->render('admin/media/index.html.twig', [
             'medias' => $medias,
@@ -50,10 +59,10 @@ class MediaController extends AbstractController
             if (!$this->isGranted('ROLE_ADMIN')) {
                 $media->setUser($this->getUser());
             }
-            $media->setPath('uploads/' . md5(uniqid()) . '.' . $media->getFile()->guessExtension());
-            $media->getFile()->move('uploads/', $media->getPath());
-            $this->getDoctrine()->getManager()->persist($media);
-            $this->getDoctrine()->getManager()->flush();
+            $media->setPath('uploads/' . md5(uniqid('', true)) . '.' . $media->getFile()?->guessExtension());
+            $media->getFile()?->move('uploads/', $media->getPath());
+            $this->entityManager->persist($media);
+            $this->entityManager->flush();
 
             return $this->redirectToRoute('admin_media_index');
         }
@@ -65,10 +74,12 @@ class MediaController extends AbstractController
 	#[Route(path: '/admin/media/delete/{id}', name: 'admin_media_delete')]
 	public function delete(int $id): RedirectResponse
 	{
-        $media = $this->getDoctrine()->getRepository(Media::class)->find($id);
-        $this->getDoctrine()->getManager()->remove($media);
-        $this->getDoctrine()->getManager()->flush();
-        unlink($media->getPath());
+        $media = $this->entityManager->getRepository(Media::class)->find($id);
+		if ($media){
+			$this->entityManager->remove($media);
+			$this->entityManager->flush();
+			unlink($media->getPath());
+		}
 
         return $this->redirectToRoute('admin_media_index');
     }
